+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All Notable changes to `League\Url` will be documented in this file
 - `isEmpty` method to `League\Url\Interfaces\Url` to tell whether a URL is empty or not
 - `isSupported` static method to `League\Url\Scheme` to tell whether a specified scheme is supported by the library
 - Improve decoding invalid punycode host labels
+- Add support for `gopher` scheme
 
 ## Fixed